Remove obsolete tooltip file
