+++ src/components/building.rs
@@ -1,0 +1,12 @@
+use bevy::prelude::*;
+
+#[derive(Component)]
+pub struct OperationsHub;
+#[derive(Component)]
+pub struct BioDome;
+#[derive(Component)]
+pub struct PowerRelay;
+#[derive(Component)]
+pub struct StorageSilo;
+#[derive(Component)]
+pub struct ResearchInstitute;
