+++ src/components/citizen.rs
@@ -1,0 +1,7 @@
+use bevy::prelude::*;
+
+#[derive(Component)]
+pub struct Specialist;
+
+#[derive(Component)]
+pub struct Inhabitant;
