+++ src/components/mod.rs
@@ -1,0 +1,2 @@
+pub mod building;
+pub mod citizen;
