+++ src/resources/mod.rs
@@ -1,0 +1,2 @@
+pub mod population;
+pub mod tutorial;
